fix: popup component
